Update bubblewrap to 0.2.1
Summary:
This is a minor release with some fixes and cleanups.

We now distribute all the demos in the tarball and there was some fixes to make them work on more distributions and with different versions of python.

There was an issue with mkdir when running bubblewrap on an NFS filesystem that has been fixed, so flatpak now works on NFS shares.

Some leaks have been fixed, including a file descriptor leak.

bubblewrap now builds on systems without PR_CAP_AMBIENT.

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan:
- run `bwrap`
- build flatpak against this version

Reviewers: #triage_team, sunnyflunk

Reviewed By: #triage_team, sunnyflunk

Differential Revision: https://dev.solus-project.com/D3217