Update bubblewrap to 0.3.1
Summary:
New feature in this release is --bind-try (as well as --dev-bind-try and --ro-bind-try) which works like the regular versions if the source exists, but does nothing if it doesn't exist.

The mount type for the root tmpfs was also changed to "tmpfs" instead of being empty, as the later could cause problems with some programs when parsing the mountinfo files in /proc.

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: Successfully build flatpak against this version

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D3999